moved example src to os_assistant
