+++ Makefile
@@ -13,5 +13,9 @@ test:
 tox:
 	uv run tox
 
+# TODO: the following clean command would also clean the virtual environment so don't use it
+clean:
+	git clean -fxfd -e '*venv*' --dry-run
+
 demo:
 	uv run examples/os_assistant.py
