--- examples/Experimental_Code_lang/src/__init__.py
@@ -1,1 +1,0 @@
-# This file marks the src directory as a Python package.
