--- examples/Experimental_Code_lang/src/models/__init__.py
@@ -1,1 +1,0 @@
-# This file marks the models directory as a Python package.
